+++ src/domain/ConsumeEventStreamRecord.ts
