+++ src/domain/Notification.ts
@@ -6,7 +6,7 @@ import * as RA from 'fp-ts/ReadonlyArray';
 import { Iun } from '../generated/definitions/Iun';
 import { FullSentNotification } from '../generated/definitions/FullSentNotification';
 import { CheckNotificationStatusRecord } from './CheckNotificationStatusRecord';
-import { ConsumeEventStreamRecord, getProgressResponseList } from './ConsumeEventStreamRecordRepository';
+import { ConsumeEventStreamRecord, getProgressResponseList } from './ConsumeEventStreamRecord';
 import { makeNotificationRequestFromFind, NotificationRequest } from './NotificationRequest';
 import { makeFullSentNotification } from './GetNotificationDetailRepository';
