+++ src/useCases/GetNotificationDetailUseCase.ts
@@ -10,7 +10,7 @@ import { authorizeApiKey } from '../domain/authorize';
 import { computeSnapshot } from '../domain/Snapshot';
 import { isNewNotificationRecord } from '../domain/NewNotificationRecord';
 import { isCheckNotificationStatusRecord } from '../domain/CheckNotificationStatusRecord';
-import { isConsumeEventStreamRecord } from '../domain/ConsumeEventStreamRecordRepository';
+import { isConsumeEventStreamRecord } from '../domain/ConsumeEventStreamRecord';
 import { SystemEnv } from './SystemEnv';
 
 // TODO: Apply the Reader monad to the environment.
