+++ src/useCases/GetPaymentNotificationMetadataUseCase.ts
@@ -15,7 +15,7 @@ import { NotificationPaymentAttachment } from '../generated/definitions/Notifica
 import { NotificationPaymentInfo } from '../generated/definitions/NotificationPaymentInfo';
 import { isNewNotificationRecord } from '../domain/NewNotificationRecord';
 import { isCheckNotificationStatusRecord } from '../domain/CheckNotificationStatusRecord';
-import { isConsumeEventStreamRecord } from '../domain/ConsumeEventStreamRecordRepository';
+import { isConsumeEventStreamRecord } from '../domain/ConsumeEventStreamRecord';
 import { SystemEnv } from './SystemEnv';
 
 // FIXME: The attachmentName type should be an enum -> check type generated by openapi
